+++ docs/Changelog.md
@@ -4,6 +4,8 @@ This section lists changes not yet released, but committed to repositories stunt
 WIP means work in progress, unfinished, still being developed.
 
 - WIP: 1 new car: LF
+- added game arguments: prv, prvfar, prvnear
+  - to start with preview camera setup for making vehicle screens
 - Updated **docs** (a lot) and screenshots
   - also added (sticky) forum topics with posts into docs/
 - [multiplayer](Multiplayer.md) added new master/game server to [game config](../config/game-default.cfg) under [ network ]